Lab vectorization 2: back to 32 bit

+++ labs/core_bound/vectorization_2/solution.cpp
@@ -1,7 +1,7 @@
 #include "solution.hpp"
 
 uint16_t checksum(const Blob &blob) {
-  uint64_t acc = 0;
+  uint32_t acc = 0;
   for (auto value : blob) {
     acc += value;
   }
